Extract report class
